CSFD-4 refactoring as per dev review. Separate class for upload_timestamp functionality. Additional test to prove field added for newly added documents.
