+++ src/contractTest/java/uk/gov/hmcts/ccd/v2/external/controller/ContractTestSubmitCaseTransaction.java
@@ -12,6 +12,7 @@ import uk.gov.hmcts.ccd.domain.service.common.SecurityClassificationService;
 import uk.gov.hmcts.ccd.domain.service.common.UIDService;
 import uk.gov.hmcts.ccd.domain.service.createcase.SubmitCaseTransaction;
 import uk.gov.hmcts.ccd.domain.service.getcasedocument.CaseDocumentService;
+import uk.gov.hmcts.ccd.domain.service.getcasedocument.CaseDocumentTimestampService;
 import uk.gov.hmcts.ccd.domain.service.message.MessageService;
 import uk.gov.hmcts.ccd.domain.service.stdapi.CallbackInvoker;
 
@@ -29,9 +30,10 @@ public class ContractTestSubmitCaseTransaction extends SubmitCaseTransaction {
                                              SecurityClassificationService securityClassificationService,
                                              CaseDataAccessControl caseDataAccessControl,
                                              MessageService messageService,
-                                             CaseDocumentService caseDocumentService) {
+                                             CaseDocumentService caseDocumentService,
+                                             CaseDocumentTimestampService caseDocumentTimestampService) {
         super(caseDetailsRepository, caseAuditEventRepository, caseTypeService,
             callbackInvoker, uidService, securityClassificationService,
-            caseDataAccessControl, messageService, caseDocumentService);
+            caseDataAccessControl, messageService, caseDocumentService, caseDocumentTimestampService);
     }
 }
