+++ src/main/java/uk/gov/hmcts/ccd/domain/service/createcase/SubmitCaseTransaction.java
@@ -22,6 +22,7 @@ import uk.gov.hmcts.ccd.domain.service.common.CaseTypeService;
 import uk.gov.hmcts.ccd.domain.service.common.SecurityClassificationService;
 import uk.gov.hmcts.ccd.domain.service.common.UIDService;
 import uk.gov.hmcts.ccd.domain.service.getcasedocument.CaseDocumentService;
+import uk.gov.hmcts.ccd.domain.service.getcasedocument.CaseDocumentTimestampService;
 import uk.gov.hmcts.ccd.domain.service.message.MessageContext;
 import uk.gov.hmcts.ccd.domain.service.message.MessageService;
 import uk.gov.hmcts.ccd.domain.service.stdapi.AboutToSubmitCallbackResponse;
@@ -46,6 +47,7 @@ public class SubmitCaseTransaction implements AccessControl {
     private final CaseDataAccessControl caseDataAccessControl;
     private final MessageService messageService;
     private final CaseDocumentService caseDocumentService;
+    private final CaseDocumentTimestampService caseDocumentTimestampService;
 
     @Inject
     public SubmitCaseTransaction(@Qualifier(CachedCaseDetailsRepository.QUALIFIER)
@@ -57,7 +59,8 @@ public class SubmitCaseTransaction implements AccessControl {
                                  final SecurityClassificationService securityClassificationService,
                                  final CaseDataAccessControl caseDataAccessControl,
                                  final @Qualifier("caseEventMessageService") MessageService messageService,
-                                 final CaseDocumentService caseDocumentService
+                                 final CaseDocumentService caseDocumentService,
+                                 final CaseDocumentTimestampService caseDocumentTimestampService
                                  ) {
         this.caseDetailsRepository = caseDetailsRepository;
         this.caseAuditEventRepository = caseAuditEventRepository;
@@ -68,6 +71,7 @@ public class SubmitCaseTransaction implements AccessControl {
         this.caseDataAccessControl = caseDataAccessControl;
         this.messageService = messageService;
         this.caseDocumentService = caseDocumentService;
+        this.caseDocumentTimestampService = caseDocumentTimestampService;
     }
 
     @Transactional(propagation = Propagation.REQUIRES_NEW)
@@ -108,7 +112,7 @@ public class SubmitCaseTransaction implements AccessControl {
             ignoreWarning
         );
 
-        caseDocumentService.addUploadTimestamps(caseDetailsWithoutHashes);
+        caseDocumentTimestampService.addUploadTimestamps(caseDetailsWithoutHashes, null);
 
         @SuppressWarnings("UnnecessaryLocalVariable")
         final CaseDetails caseDetailsAfterCallback = caseDetailsWithoutHashes;
