+++ src/test/java/uk/gov/hmcts/ccd/domain/service/createcase/SubmitCaseTransactionTest.java
@@ -28,6 +28,7 @@ import uk.gov.hmcts.ccd.domain.service.common.CaseTypeService;
 import uk.gov.hmcts.ccd.domain.service.common.SecurityClassificationServiceImpl;
 import uk.gov.hmcts.ccd.domain.service.common.UIDService;
 import uk.gov.hmcts.ccd.domain.service.getcasedocument.CaseDocumentService;
+import uk.gov.hmcts.ccd.domain.service.getcasedocument.CaseDocumentTimestampService;
 import uk.gov.hmcts.ccd.domain.service.message.MessageContext;
 import uk.gov.hmcts.ccd.domain.service.message.MessageService;
 import uk.gov.hmcts.ccd.domain.service.stdapi.AboutToSubmitCallbackResponse;
@@ -82,7 +83,6 @@ class SubmitCaseTransactionTest {
     private static final String ON_BEHALF_OF_FNAME = "Pierre OnBehalf";
     private static final String ON_BEHALF_OF_LNAME = "Martin OnBehalf";
 
-
     @Mock
     private CaseDetailsRepository caseDetailsRepository;
     @Mock
@@ -110,6 +110,9 @@ class SubmitCaseTransactionTest {
     private CaseDocumentService caseDocumentService;
 
     @Mock
+    private CaseDocumentTimestampService caseDocumentTimestampService;
+
+    @Mock
     private MessageService messageService;
 
     @InjectMocks
@@ -132,7 +135,8 @@ class SubmitCaseTransactionTest {
             securityClassificationService,
             caseDataAccessControl,
             messageService,
-            caseDocumentService
+            caseDocumentService,
+            caseDocumentTimestampService
         );
 
         event = buildEvent();
@@ -273,7 +277,6 @@ class SubmitCaseTransactionTest {
         );
     }
 
-
     @Test
     @DisplayName("should invoke callback")
     void shouldInvokeCallback() {
@@ -395,7 +398,6 @@ class SubmitCaseTransactionTest {
         event.setName(EVENT_NAME);
         return event;
     }
-
 
     private IdamUser buildIdamUser() {
         final IdamUser idamUser = new IdamUser();
