+++ src/test/java/uk/gov/hmcts/ccd/domain/service/getcasedocument/CaseDocumentServiceTest.java
@@ -1,12 +1,7 @@
 package uk.gov.hmcts.ccd.domain.service.getcasedocument;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.JsonMappingException;
 import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.JsonNodeFactory;
-import com.google.common.collect.Lists;
-import com.google.common.collect.Maps;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -24,9 +19,6 @@ import uk.gov.hmcts.ccd.endpoint.exceptions.ServiceException;
 import uk.gov.hmcts.ccd.endpoint.exceptions.ValidationException;
 import uk.gov.hmcts.ccd.v2.external.domain.DocumentHashToken;
 
-import java.time.Instant;
-import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
@@ -38,10 +30,6 @@ import static java.util.Collections.emptySet;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.ThrowableAssert.catchThrowable;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertNotEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyList;
 import static org.mockito.ArgumentMatchers.anyMap;
@@ -52,7 +40,6 @@ import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoMoreInteractions;
-import static uk.gov.hmcts.ccd.domain.service.getcasedocument.CaseDocumentUtils.UPLOAD_TIMESTAMP;
 
 @ExtendWith(MockitoExtension.class)
 class CaseDocumentServiceTest extends TestFixtures {
@@ -305,370 +292,5 @@ class CaseDocumentServiceTest extends TestFixtures {
 
         verifyNoMoreInteractions(documentUtils);
     }
-
-    @Test
-    void testFindUrlsNotInOriginal() {
-        List<String> dbUrls = List.of(urlGoogle, urlMicrosoft, urlYahoo);
-        List<String> requestUrls = List.of(urlGoogle, urlMicrosoft, urlYahoo, urlElastic, urlApple);
-        List<String> urlResults = underTest.findUrlsNotInOriginal(dbUrls, requestUrls);
-
-        assertEquals(2, urlResults.size());
-        assertTrue(urlResults.contains(urlApple));
-        assertTrue(urlResults.contains(urlElastic));
-    }
-
-    @Test
-    void testFindAllUrlsNotInOriginal() {
-        List<String> requestUrls = List.of(urlGoogle, urlMicrosoft, urlYahoo, urlElastic, urlApple);
-        List<String> dbUrls = new ArrayList<>();
-        List<String> urlResults = underTest.findUrlsNotInOriginal(dbUrls, requestUrls);
-
-        assertEquals(5, urlResults.size());
-        assertTrue(urlResults.contains(urlApple));
-        assertTrue(urlResults.contains(urlElastic));
-        assertTrue(urlResults.contains(urlGoogle));
-        assertTrue(urlResults.contains(urlMicrosoft));
-        assertTrue(urlResults.contains(urlYahoo));
-    }
-
-    @Test
-    void testFindNoUrlsInOriginal() {
-        List<String> requestUrls = new ArrayList<>();
-        List<String> dbUrls = List.of(urlGoogle, urlMicrosoft, urlYahoo, urlElastic, urlApple);
-        List<String> urlResults = underTest.findUrlsNotInOriginal(dbUrls, requestUrls);
-
-        assertEquals(0, urlResults.size());
-    }
-
-    @Test
-    void testFindDocuments() {
-        Map<String, JsonNode> dataMap = new HashMap<>();
-        JsonNode result = generateTestNode(jsonString);
-        dataMap.put("testNode", result);
-
-        List<JsonNode> listDocuments = underTest.findNodes(dataMap.values());
-
-        assertTrue(listDocuments.size() > 0);
-        assertEquals(18, listDocuments.size());
-
-        listDocuments.forEach(e -> {
-            System.out.println(e.get("document_url"));
-        });
-    }
-
-    //    @Test
-    //    void testAddTimestamp() {
-    //        Map<String, JsonNode> dataMapOriginal = new HashMap<>();
-    //        JsonNode resultOriginal = generateTestNode(jsonStringOriginal);
-    //        dataMapOriginal.put("testNode", resultOriginal);
-    //        CaseDetails caseDetailsDb = new CaseDetails();
-    //        caseDetailsDb.setReference(Long.valueOf(1));
-    //        caseDetailsDb.setData(dataMapOriginal);
-    //
-    //        when(caseService.getCaseDetailsByCaseReference(any())).thenReturn(caseDetailsDb);
-    //
-    //        Map<String, JsonNode> dataMap = new HashMap<>();
-    //        JsonNode result = generateTestNode(jsonString);
-    //        dataMapOriginal.put("testNode", result);
-    //        CaseDetails caseDetails = new CaseDetails();
-    //        caseDetailsDb.setReference(caseDetailsDb.getReference());
-    //        caseDetailsDb.setData(dataMap);
-    //
-    //        underTest.addUploadTimestamps(caseDetails);
-    //
-    //        assertTrue(caseDetails.getData().isEmpty());
-    //
-    //    }
-
-    @Test
-    void testInsertTimestampIfNotAlreadyPresent() {
-        JsonNode jsonNode = generateTestNode(jsonDocumentNode);
-        String uploadTimestamp = Instant.now().toString();
-        assertFalse(jsonNode.has(UPLOAD_TIMESTAMP));
-
-        underTest.insertUploadTimestamp(jsonNode, uploadTimestamp);
-        assertTrue(jsonNode.has(UPLOAD_TIMESTAMP));
-        assertEquals(uploadTimestamp, jsonNode.get(UPLOAD_TIMESTAMP).textValue());
-        System.out.println(jsonNode.asText());
-    }
-
-    @Test
-    void testDoNotInsertTimestampIfAlreadyPresent() {
-        final String uploadTimestamp = Instant.now().minusNanos(5).toString();
-        JsonNode jsonNode = generateTestNode(jsonDocumentNode);
-        assertFalse(jsonNode.has(UPLOAD_TIMESTAMP));
-
-        underTest.insertUploadTimestamp(jsonNode, uploadTimestamp);
-        assertTrue(jsonNode.has(UPLOAD_TIMESTAMP));
-        assertEquals(uploadTimestamp, jsonNode.get(UPLOAD_TIMESTAMP).textValue());
-
-        String uploadTimestampNew = Instant.now().toString();
-        underTest.insertUploadTimestamp(jsonNode, uploadTimestampNew);
-        assertNotEquals(uploadTimestampNew, jsonNode.get(UPLOAD_TIMESTAMP).textValue());
-        assertEquals(uploadTimestamp, jsonNode.get(UPLOAD_TIMESTAMP).textValue());
-
-        System.out.println(jsonNode.asText());
-    }
-
-    @Test
-    void testFindNewDocuments() {
-        List<String> listUrlsDb = generateListOfUrls(jsonStringOriginal);
-        assertEquals(13, listUrlsDb.size());
-
-        List<String> listUrlsRequest = generateListOfUrls(jsonString);
-        assertEquals(18, listUrlsRequest.size());
-
-        List<String> listUrlsNew = underTest.findUrlsNotInOriginal(listUrlsDb, listUrlsRequest);
-
-        assertTrue(listUrlsNew.size() > 0);
-        assertEquals(5, listUrlsNew.size());
-
-        listUrlsNew.forEach(e -> {
-            System.out.println(e);
-        });
-    }
-
-    private List<String> generateListOfUrls(String jsonString) {
-        JsonNode result = generateTestNode(jsonString);
-        Map<String, JsonNode> dataMap = Maps.newHashMap();
-        dataMap.put("testNode", result);
-        return underTest.findDocumentUrls(dataMap.values());
-    }
-
-
-    private Map<String, JsonNode> buildData(String... dataFieldIds) {
-        Map<String, JsonNode> dataMap = Maps.newHashMap();
-        Lists.newArrayList(dataFieldIds)
-            .forEach(dataFieldId -> dataMap.put(dataFieldId, JSON_NODE_FACTORY.textNode(dataFieldId)));
-        return dataMap;
-    }
-
-    private JsonNode generateTestNode(String json) {
-
-        // Create ObjectMapper
-        ObjectMapper objectMapper = new ObjectMapper();
-
-        JsonNode jsonNode = null;
-
-        try {
-            // Parse JSON string to JsonNode
-            jsonNode = objectMapper.readTree(json);
-
-        } catch (JsonMappingException e) {
-            throw new RuntimeException(e);
-        } catch (JsonProcessingException e) {
-            throw new RuntimeException(e);
-        }
-        return jsonNode;
-
-    }
-
-    private static String jsonDocumentNode = "{\n"
-        + "  \"document\": {\n"
-        + "     \"document_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/63122c23-3665-4dd1-8f81-03d0cb86cac4\",\n"
-        + "     \"document_filename\": \"PD36Q letter.pdf\",\n"
-        + "     \"document_binary_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/63122c23-3665-4dd1-8f81-03d0cb86cac4/binary\"\n"
-        + "   }\n"
-        + "}";
-
-    private static String jsonString = "{\n"
-        + "    \"id\": \"1675936805799936\",\n"
-        + "    \"additionalApplicationsBundle\": [\n"
-        + "        {\n"
-        + "            \"id\": \"6f5418ac-e59a-42f6-84d0-a9d97c519a4a\",\n"
-        + "            \"uploadedDateTime\": \"27-Feb-2023 09:44:18 am\",\n"
-        + "            \"otherApplicationsBundle\": {\n"
-        + "                \"author\": \"prl_aat_solicitor@mailinator.com\",\n"
-        + "                \"document\": {\n"
-        + "                    \"document_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/63122c23-3665-4dd1-8f81-03d0cb86cac4\",\n"
-        + "                    \"document_filename\": \"PD36Q letter.pdf\",\n"
-        + "                    \"document_binary_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/63122c23-3665-4dd1-8f81-03d0cb86cac4/binary\",\n"
-        + "                    \"upload_timestamp\": \"2023-03-01T12:34:56\"\n"
-        + "                },\n"
-        + "                \"newDocument4\": {\n"
-        + "                    \"document_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/c50985d5-04cd-4f36-a9db-630bcd4c1aaa\",\n"
-        + "                    \"document_filename\": \"NewDoc4.pdf\",\n"
-        + "                    \"document_binary_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/c50985d5-04cd-4f36-a9db-630bcd4c1aaa/binary\",\n"
-        + "                    \"upload_timestamp\": \"2024-01-11T17:22:30\"\n"
-        + "                }\n"
-        + "            }\n"
-        + "        }\n"
-        + "    ],\n"
-        + "   \"orderCollection\": "
-        + "   [\n"
-        + "   {\n"
-        + "       \"id\": \"14cadd3a-1afd-46c1-8805-bd16bdcdb489\",\n"
-        + "       \"orderDocument\": {\n"
-        + "       \"document_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/f94d7ea6-fbed-4b5d-8155-4c4851c277c8\",\n"
-        + "       \"document_filename\": \"Welsh_ChildArrangements_Specific_Prohibited_Steps_C43.pdf\",\n"
-        + "       \"document_binary_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/f94d7ea6-fbed-4b5d-8155-4c4851c277c8/binary\"\n"
-        + "       },\n"
-        + "       \"newDocument5\": {\n"
-        + "       \"document_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/c50985d5-04cd-4f36-a9db-630bcd4c1bbb\",\n"
-        + "       \"document_filename\": \"NewDoc5.pdf\",\n"
-        + "       \"document_binary_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/c50985d5-04cd-4f36-a9db-630bcd4c1bbb/binary\",\n"
-        + "       \"upload_timestamp\": \"2024-01-11T17:22:30\"\n"
-        + "       }\n"
-        + "   }\n"
-        + "   ],\n"
-        + "\"previewOrderDoc\": {\n"
-        + "   \"document_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/20555d12-2ee7-4cf3-b827-3a0d9f13753c\",\n"
-        + "   \"document_filename\": \"ChildArrangements_Specific_Prohibited_Steps_C43_Draft.pdf\",\n"
-        + "   \"document_binary_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/20555d12-2ee7-4cf3-b827-3a0d9f13753c/binary\"\n"
-        + "},\n"
-        + "\"finalWelshDocument\": {\n"
-        + "   \"document_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/d6172765-31b2-4985-bbf5-e70ff3280459\",\n"
-        + "   \"document_filename\": \"C100FinalDocumentWelsh.pdf\",\n"
-        + "   \"document_binary_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/d6172765-31b2-4985-bbf5-e70ff3280459/binary\"\n"
-        + "},\n"
-        + "\"submitAndPayDownloadApplicationLink\": {\n"
-        + "   \"document_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/31d4c664-c8a8-447e-a5ea-57c3594ee78e\",\n"
-        + "   \"document_filename\": \"Draft_C100_application.pdf\",\n"
-        + "   \"document_binary_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/31d4c664-c8a8-447e-a5ea-57c3594ee78e/binary\"\n"
-        + "},\n"
-        + "\"draftConsentOrderFile\": {\n"
-        + "   \"document_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/ebefdc86-b523-474e-a3bc-c06a931f1174\",\n"
-        + "   \"document_filename\": \"Draft consent order - Smith.pdf\",\n"
-        + "   \"document_binary_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/ebefdc86-b523-474e-a3bc-c06a931f1174/binary\"\n"
-        + "},\n"
-        + "\"c8WelshDocument\": {\n"
-        + "   \"document_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/40101fec-d005-47a6-b9e9-13c6a97f1dae\",\n"
-        + "   \"document_filename\": \"C8Document_Welsh.pdf\",\n"
-        + "   \"document_binary_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/40101fec-d005-47a6-b9e9-13c6a97f1dae/binary\"\n"
-        + "},\n"
-        + "\"newDocument3\": {\n"
-        + "   \"document_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/c50985d5-04cd-4f36-a9db-630bcd4c1zzz\",\n"
-        + "   \"document_filename\": \"NewDoc3.pdf\",\n"
-        + "   \"document_binary_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/c50985d5-04cd-4f36-a9db-630bcd4c1yyy/zzz\",\n"
-        + "   \"upload_timestamp\": \"2024-01-11T17:22:30\"\n"
-        + "},\n"
-        + "\"previewOrderDocWelsh\": {\n"
-        + "   \"document_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/01f852b8-edaa-488e-8463-2acefad20c87\",\n"
-        + "   \"document_filename\": \"Welsh_ChildArrangements_Specific_Prohibited_Steps_C43_Draft.pdf\",\n"
-        + "   \"document_binary_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/01f852b8-edaa-488e-8463-2acefad20c87/binary\"\n"
-        + "},\n"
-        + "\"c1AWelshDocument\": {\n"
-        + "   \"document_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/e81c1756-d7c1-4c9a-8d0d-45d96a13b417\",\n"
-        + "   \"document_filename\": \"C1A_Document_Welsh.pdf\",\n"
-        + "   \"document_binary_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/e81c1756-d7c1-4c9a-8d0d-45d96a13b417/binary\"\n"
-        + "},\n"
-        + "\"c8Document\": {\n"
-        + "   \"document_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/c2300009-51ec-405d-9770-df67ec4e6bc2\",\n"
-        + "   \"document_filename\": \"C8Document.pdf\",\n"
-        + "   \"document_binary_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/c2300009-51ec-405d-9770-df67ec4e6bc2/binary\"\n"
-        + "},\n"
-        + "\"finalDocument\": {\n"
-        + "   \"document_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/2cfd7aff-37f6-4b6d-9104-770331d0ee06\",\n"
-        + "   \"document_filename\": \"C100FinalDocument.pdf\",\n"
-        + "   \"document_binary_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/2cfd7aff-37f6-4b6d-9104-770331d0ee06/binary\"\n"
-        + "},\n"
-        + "\"submitAndPayDownloadApplicationWelshLink\": {\n"
-        + "   \"document_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/c50985d5-04cd-4f36-a9db-630bcd4c1b6b\",\n"
-        + "   \"document_filename\": \"Draft_C100_application_welsh.pdf\",\n"
-        + "   \"document_binary_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/c50985d5-04cd-4f36-a9db-630bcd4c1b6b/binary\"\n"
-        + "},\n"
-        + "\"newDocument1\": {\n"
-        + "   \"document_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/c50985d5-04cd-4f36-a9db-630bcd4c1xxx\",\n"
-        + "   \"document_filename\": \"NewDoc1.pdf\",\n"
-        + "   \"document_binary_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/c50985d5-04cd-4f36-a9db-630bcd4c1xxx/binary\",\n"
-        + "   \"upload_timestamp\": \"2024-01-11T17:22:30\"\n"
-        + "},\n"
-        + "\"newDocument2\": {\n"
-        + "   \"document_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/c50985d5-04cd-4f36-a9db-630bcd4c1yyy\",\n"
-        + "   \"document_filename\": \"NewDoc2.pdf\",\n"
-        + "   \"document_binary_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/c50985d5-04cd-4f36-a9db-630bcd4c1yyy/binary\",\n"
-        + "   \"upload_timestamp\": \"2024-01-11T17:22:30\"\n"
-        + "},\n"
-        + "\"c1ADocument\": {\n"
-        + "   \"document_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/ca1f2ba2-21a6-4b75-9c6f-3f1c5b5564be\",\n"
-        + "   \"document_filename\": \"C1A_Document.pdf\",\n"
-        + "   \"document_binary_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/ca1f2ba2-21a6-4b75-9c6f-3f1c5b5564be/binary\"\n"
-        + "}\n"
-        + "}";
-
-    private static String jsonStringOriginal = "{\n"
-        + "    \"id\": \"1675936805799936\",\n"
-        + "    \"additionalApplicationsBundle\": [\n"
-        + "        {\n"
-        + "            \"id\": \"6f5418ac-e59a-42f6-84d0-a9d97c519a4a\",\n"
-        + "            \"uploadedDateTime\": \"27-Feb-2023 09:44:18 am\",\n"
-        + "            \"otherApplicationsBundle\": {\n"
-        + "                \"author\": \"prl_aat_solicitor@mailinator.com\",\n"
-        + "                \"document\": {\n"
-        + "                    \"document_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/63122c23-3665-4dd1-8f81-03d0cb86cac4\",\n"
-        + "                    \"document_filename\": \"PD36Q letter.pdf\",\n"
-        + "                    \"document_binary_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/63122c23-3665-4dd1-8f81-03d0cb86cac4/binary\",\n"
-        + "                    \"upload_timestamp\": \"2023-03-01T12:34:56\"\n"
-        + "                }\n"
-        + "            }\n"
-        + "        }\n"
-        + "    ],\n"
-        + "   \"orderCollection\": "
-        + "   [\n"
-        + "   {\n"
-        + "       \"id\": \"14cadd3a-1afd-46c1-8805-bd16bdcdb489\",\n"
-        + "       \"orderDocument\": {\n"
-        + "       \"document_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/f94d7ea6-fbed-4b5d-8155-4c4851c277c8\",\n"
-        + "       \"document_filename\": \"Welsh_ChildArrangements_Specific_Prohibited_Steps_C43.pdf\",\n"
-        + "       \"document_binary_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/f94d7ea6-fbed-4b5d-8155-4c4851c277c8/binary\"\n"
-        + "       }\n"
-        + "   }\n"
-        + "   ],\n"
-        + "\"previewOrderDoc\": {\n"
-        + "   \"document_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/20555d12-2ee7-4cf3-b827-3a0d9f13753c\",\n"
-        + "   \"document_filename\": \"ChildArrangements_Specific_Prohibited_Steps_C43_Draft.pdf\",\n"
-        + "   \"document_binary_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/20555d12-2ee7-4cf3-b827-3a0d9f13753c/binary\"\n"
-        + "},\n"
-        + "\"finalWelshDocument\": {\n"
-        + "   \"document_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/d6172765-31b2-4985-bbf5-e70ff3280459\",\n"
-        + "   \"document_filename\": \"C100FinalDocumentWelsh.pdf\",\n"
-        + "   \"document_binary_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/d6172765-31b2-4985-bbf5-e70ff3280459/binary\"\n"
-        + "},\n"
-        + "\"submitAndPayDownloadApplicationLink\": {\n"
-        + "   \"document_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/31d4c664-c8a8-447e-a5ea-57c3594ee78e\",\n"
-        + "   \"document_filename\": \"Draft_C100_application.pdf\",\n"
-        + "   \"document_binary_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/31d4c664-c8a8-447e-a5ea-57c3594ee78e/binary\"\n"
-        + "},\n"
-        + "\"draftConsentOrderFile\": {\n"
-        + "   \"document_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/ebefdc86-b523-474e-a3bc-c06a931f1174\",\n"
-        + "   \"document_filename\": \"Draft consent order - Smith.pdf\",\n"
-        + "   \"document_binary_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/ebefdc86-b523-474e-a3bc-c06a931f1174/binary\"\n"
-        + "},\n"
-        + "\"c8WelshDocument\": {\n"
-        + "   \"document_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/40101fec-d005-47a6-b9e9-13c6a97f1dae\",\n"
-        + "   \"document_filename\": \"C8Document_Welsh.pdf\",\n"
-        + "   \"document_binary_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/40101fec-d005-47a6-b9e9-13c6a97f1dae/binary\"\n"
-        + "},\n"
-        + "\"previewOrderDocWelsh\": {\n"
-        + "   \"document_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/01f852b8-edaa-488e-8463-2acefad20c87\",\n"
-        + "   \"document_filename\": \"Welsh_ChildArrangements_Specific_Prohibited_Steps_C43_Draft.pdf\",\n"
-        + "   \"document_binary_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/01f852b8-edaa-488e-8463-2acefad20c87/binary\"\n"
-        + "},\n"
-        + "\"c1AWelshDocument\": {\n"
-        + "   \"document_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/e81c1756-d7c1-4c9a-8d0d-45d96a13b417\",\n"
-        + "   \"document_filename\": \"C1A_Document_Welsh.pdf\",\n"
-        + "   \"document_binary_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/e81c1756-d7c1-4c9a-8d0d-45d96a13b417/binary\"\n"
-        + "},\n"
-        + "\"c8Document\": {\n"
-        + "   \"document_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/c2300009-51ec-405d-9770-df67ec4e6bc2\",\n"
-        + "   \"document_filename\": \"C8Document.pdf\",\n"
-        + "   \"document_binary_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/c2300009-51ec-405d-9770-df67ec4e6bc2/binary\"\n"
-        + "},\n"
-        + "\"finalDocument\": {\n"
-        + "   \"document_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/2cfd7aff-37f6-4b6d-9104-770331d0ee06\",\n"
-        + "   \"document_filename\": \"C100FinalDocument.pdf\",\n"
-        + "   \"document_binary_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/2cfd7aff-37f6-4b6d-9104-770331d0ee06/binary\"\n"
-        + "},\n"
-        + "\"submitAndPayDownloadApplicationWelshLink\": {\n"
-        + "   \"document_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/c50985d5-04cd-4f36-a9db-630bcd4c1b6b\",\n"
-        + "   \"document_filename\": \"Draft_C100_application_welsh.pdf\",\n"
-        + "   \"document_binary_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/c50985d5-04cd-4f36-a9db-630bcd4c1b6b/binary\"\n"
-        + "},\n"
-        + "\"c1ADocument\": {\n"
-        + "   \"document_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/ca1f2ba2-21a6-4b75-9c6f-3f1c5b5564be\",\n"
-        + "   \"document_filename\": \"C1A_Document.pdf\",\n"
-        + "   \"document_binary_url\": \"http://dm-store-aat.service.core-compute-aat.internal/documents/ca1f2ba2-21a6-4b75-9c6f-3f1c5b5564be/binary\"\n"
-        + "}\n"
-        + "}";
 
 }
